pyproject.toml
[project]
name = "workspace"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # "django-ninja>=1.4.5,<1.5",
    "django-ninja==1.5.0b1",
    "rozklad-ontu-parser-makisukurisu>=0.1.2.0,<0.2",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
extend-select = [
  "UP",
  "E",
  "W",
  "I",
  "C90",
  "N",
  "ANN",
  "ASYNC",
  "BLE",
  "FBT",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "FA",
  "ISC",
  "ICN",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "Q",
  "RET",
  "SLF",
  "SIM",
  "TID",
  "TCH",
  "ARG",
  "TD",
  "FIX",
  "ERA",
  "PL",
  "RUF",
]
ignore = [
  "ISC001",
  "N999",
  "ERA001",
  "RUF012",
  "FBT001",
  "FBT002",
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
